ao_wasapi: remove always triggered messages
Depending on the log level, this will lock the audio thread if any other
thread is writing a message.

These messages can be restored for debugging if needed, but they do not
provide much value on their own. If anything, they should be printed
only if the values are outside the expected range, not on every frame.