options: add --osd-show-playlist-titles
Allow configuring whether to print the media-title or the filename in
show-text ${playlist}, the OSC playlist and in the playlist selector.

Showing titles can be annoying when files are badly tagged, or when it
hides the track numbers of album songs.

The OSC's playlist_media_title script-opt is removed because this option
is better since it works everywhere after configuring it once.

Closes #11653.