options: extend --osd-playlist-style to show both title and filename
This prints `<title> (<filename>)` if set to `both` mode. Depending on
the playback context, it may be more important to show the title or
filename. While the option to select one or the other helps, it is not
convenient to have to switch it each time. This commit adds an option to
show both.