ci: refactor ci so common args are picked from one location
Mostly. It is annoying if we want to add some flag/option to all the cis
and then have to do it manually. Things easily get missed that way.
Skipped powershell because of course I'm not going to deal with that.
The other problem is that mingw seems to have some mysterious dependence
on the order of cflags and trying to use them from the common place
causes the build to fail due to missing printf attributes. Give up and
don't bother with compiler flags but at least we can use other meson
arguments without an issue.