player/video: reset audio sync speed to 0 when below the threshold
Currently audio speed is adjusted only when av diff is above the
threshold. This can cause oscillation when big speed offset is applied,
which causes it to overshoot the target, especially because if we enter
the ranged close to 0 av_diff (perfect sync) we still have possibly big
audio speed factor applied, which is never adjusted untill it goes out
of sync again.

To preserve initial idea of not adjusting audio speed on every frame,
reset its speed when close to perfect sync below the vsync / 2 diff.
This will allow it keep sync or slowly drift and once above the
threshold adjust the speed again.

Fixes oscillation and constant audio speed adjustment, while it was
completely not necessary.

Fixes: c362c3d7ae85ac65e5e87004d33775f3cb291498